build: parallelize build
While building PX4Firmware for ardupilot we see several messages like:

	make[1]: warning: jobserver unavailable: using -j1.  Add '+' to parent make rule.
	make[1]: Entering directory '/home/lucas/p/dronecode/ardupilot'

The problem is that we are recursing the makefiles by simply calling
make and we don't let Make keep track of the jobs.

There are some rules that need to be serialized, but they are not in
PX4Firmware, it's Ardupilot that needs to serialize some of the
recursive makes.